(refactor:adminMemberService)
서비스 단을 새로 만들기보단 기존의 멤버에서 사용하던 서비스단에 api를 추가하는 형식으로 바꿈
