+++ src/main/java/com/easytrip/backend/admin/service/AdminService.java
@@ -2,6 +2,8 @@ package com.easytrip.backend.admin.service;
 
 import com.easytrip.backend.admin.dto.MemberDetailDto;
 import com.easytrip.backend.member.dto.request.UpdateRequest;
+import com.easytrip.backend.place.dto.PlaceDto;
+import com.easytrip.backend.place.dto.request.PlaceRequest;
 import com.easytrip.backend.type.MemberStatus;
 import com.easytrip.backend.type.SearchOption;
 import java.util.List;
