+++ src/main/java/com/easytrip/backend/member/service/ManagementService.java
@@ -1,5 +1,6 @@
 package com.easytrip.backend.member.service;
 
+import com.easytrip.backend.admin.dto.MemberDetailDto;
 import com.easytrip.backend.member.domain.MemberEntity;
 import com.easytrip.backend.member.dto.MemberDto;
 import com.easytrip.backend.member.dto.TokenCreateDto;
@@ -7,7 +8,10 @@ import com.easytrip.backend.member.dto.request.LoginRequest;
 import com.easytrip.backend.member.dto.request.ResetRequest;
 import com.easytrip.backend.member.dto.request.SignUpRequest;
 import com.easytrip.backend.member.dto.request.UpdateRequest;
+import com.easytrip.backend.type.MemberStatus;
 import com.easytrip.backend.type.Platform;
+import com.easytrip.backend.type.SearchOption;
+import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -35,4 +39,13 @@ public interface ManagementService {
   MemberDto myInfo(String accessToken);
 
   MemberDto update(String accessToken, UpdateRequest updateRequest, MultipartFile file);
+
+  void setMemberStatus(String accessToken, Long memberId, MemberStatus memberStatus);
+
+  MemberDetailDto getMemberInfo(String accessToken, Long memberId);
+
+  MemberDetailDto updateMemberInfo(String accessToken, Long memberId, UpdateRequest updateRequest,
+      MultipartFile file);
+
+  List<MemberDetailDto> searchMember(String accessToken, String keyword, SearchOption searchOption);
 }
